refactor: consolidate isRecordOfString into shared/guards.ts; drop unused replaceAllSnippets
`isRecordOfString` existed in two places with identical implementations:
`src/store/schema.ts` (consumed by `store/snippets.ts` via the unused
`replaceAllSnippets`) and `src/services/utils.ts` (consumed by `BasicTab`
JSON import). Drift risk on a security-adjacent guard.

Consolidated to `src/shared/guards.ts` per architect-review's L-002
recommendation (cross-cutting type guards belong in `shared/`, not
`services/`). One implementation, one test suite, no semantic change.

Also removed `replaceAllSnippets` from `store/snippets.ts` — it was the
only consumer of the `store/schema.ts` copy and itself had no callers in
production code (`grep -rn` only finds the test).

Changes:
- New: `src/shared/guards.ts` + `src/shared/guards.test.ts`
- Renamed: `src/store/schema.ts` → `src/shared/guards.ts` (git rename detection)
- Removed: `src/store/schema.test.ts` (tests folded into `guards.test.ts`)
- Removed: `isRecordOfString` block in `src/services/utils.ts` + its test in `utils.test.ts`
- Removed: `replaceAllSnippets` in `src/store/snippets.ts` + its test
- Updated imports: `store/snippets.ts` and `ui/components/BasicTab.ts`

Closes backlog item B-016. Closes one of the "wrong-way" arrows from
architect L-002 (store → services for a type guard). No behavior change.
