+++ src/services/utils.ts
@@ -21,13 +21,7 @@ export function diffIncoming(
     return { added, conflicts };
 }
 
-export function isRecordOfString(x: unknown): x is Record<string, string> {
-    if (!x || typeof x !== "object") return false;
-    for (const [k, v] of Object.entries(x as Record<string, unknown>)) {
-        if (typeof k !== "string" || typeof v !== "string") return false;
-    }
-    return true;
-}
+// `isRecordOfString` moved to `src/shared/guards.ts` in 1.0.9.
 
 export function normalizeTrigger(raw: string): string {
     return raw.trim();
