+++ src/store/snippets.test.ts
@@ -2,7 +2,6 @@ import { describe, it, expect } from "vitest";
 import {
     getDict,
     mergeDefaults,
-    replaceAllSnippets,
     getAllSnippetsFlat,
     hasTriggerCollision,
     hasReplacementCollision
@@ -29,15 +28,6 @@ describe("store/snippets", () => {
         const defs = { a: "A", b: "B" };
         const cur = { b: "b-cur", c: "C" };
         expect(mergeDefaults(cur, defs)).toEqual({ a: "A", b: "b-cur", c: "C" });
-    });
-
-    it("replaceAllSnippets accepts valid input and rejects invalid", () => {
-        const settings = { snippets: { a: "1" } } as any;
-        const ok = replaceAllSnippets(settings, { x: "X" });
-        expect(ok).toEqual({ snippets: { x: "X" } });
-
-        const bad = replaceAllSnippets(settings, { x: 1 } as any);
-        expect(bad).toBeNull();
     });
 
     it("hasTriggerCollision detects the same trigger name across groups", () => {
